# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMANTIC - Change & Configuration API",
        "description": "API untuk CMDB, Change Management, Maintenance & Patch Management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/reports/change-summary": {
            "get": {
                "tags": [
                    "Reports & Audit"
                ],
                "summary": "Laporan ringkasan Change Management (RFC Summary)",
                "description": "Menghasilkan ringkasan statistik dari seluruh RFC (Request for Change), mencakup status, kategori, dan jumlah total RFC per OPD.",
                "operationId": "883067fc0bc8d7b23966fa71952e9689",
                "responses": {
                    "200": {
                        "description": "Ringkasan statistik Change Management",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "overview": {
                                            "properties": {
                                                "total_rfc": {
                                                    "type": "integer",
                                                    "example": 124
                                                },
                                                "approved": {
                                                    "type": "integer",
                                                    "example": 87
                                                },
                                                "rejected": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "in_progress": {
                                                    "type": "integer",
                                                    "example": 25
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "by_category": {
                                            "properties": {
                                                "normal": {
                                                    "type": "integer",
                                                    "example": 72
                                                },
                                                "standard": {
                                                    "type": "integer",
                                                    "example": 40
                                                },
                                                "emergency": {
                                                    "type": "integer",
                                                    "example": 12
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "chart": {
                                            "properties": {
                                                "labels": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "approved"
                                                    }
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": 30
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reports/patch-compliance": {
            "get": {
                "tags": [
                    "Reports & Audit"
                ],
                "summary": "Laporan kepatuhan Patch Deployment (Patch Compliance)",
                "description": "Menghasilkan laporan kepatuhan patching berdasarkan status Patch Deployment di seluruh Configuration Item (CI).",
                "operationId": "b82ca56b1533ed53129d4ce2635cb6d2",
                "responses": {
                    "200": {
                        "description": "Ringkasan kepatuhan patch deployment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "summary": {
                                            "properties": {
                                                "total_patch_jobs": {
                                                    "type": "integer",
                                                    "example": 56
                                                },
                                                "success": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "failed": {
                                                    "type": "integer",
                                                    "example": 6
                                                },
                                                "rolled_back": {
                                                    "type": "integer",
                                                    "example": 8
                                                },
                                                "compliance_rate": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 87.5
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "by_status": {
                                            "properties": {
                                                "planned": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "deployed": {
                                                    "type": "integer",
                                                    "example": 40
                                                },
                                                "failed": {
                                                    "type": "integer",
                                                    "example": 6
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "chart": {
                                            "properties": {
                                                "labels": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Deployed"
                                                    }
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": 40
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}